% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inputs}
\alias{check_inputs}
\title{Check Function Inputs}
\usage{
check_inputs(catch_inputs)
}
\value{
Options to be passed to various \code{tmle_*} functions.
}
\description{
Check the input values of function parameters for errors.
}
\details{
...
...
}
