% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fdr_msa}
\alias{fdr_msa}
\title{FDR-MSA correction}
\usage{
fdr_msa(pvals, total_obs)
}
\arguments{
\item{pvals}{Numeric vector containing the p-values that result from any
chosen statistical hypothesis testing procedure.}

\item{total_obs}{Numeric indicating the total number of observations that
would have been available for testing prior to the selection procedure
employed in the multi-stage analysis performed.}
}
\description{
Modified FDR Controlling Procedure for Multi-Stage Analyses (MJ van der Laan
and C Tuglus, 2009, <doi:10.2202/1544-6115.1397>)
}
