% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma.R
\name{limma_screen}
\alias{limma_screen}
\title{Screening procedure based on LIMMA}
\usage{
limma_screen(methytmle, var_int, type, cutoff = 0.05)
}
\arguments{
\item{methytmle}{An object of class \code{methytmle}.}

\item{var_int}{A \code{numeric} vector containing subject-level measurements
of the variable of interest. The length of this vector must match the
number of subjects exactly. If argument \code{vim} is set to "ate" or
"rr", then the variable of interest is treated as an exposure, and the
variable must be binary in such cases. If setting \code{vim} to target
parameters assessing continuous treatment effects, then the variable
need not be binary of course.}

\item{type}{Character indicating the particular measure of DNA methylation to
be used as the observed data in the estimation procedure, either Beta
values or M-values. The data are accessed via \code{minfi::getBeta} or
\code{minfi::getM}.}

\item{cutoff}{Numeric indicating the p-value cutoff that defines which
sites pass through the \code{filter}.}
}
\value{
An object of class \code{methytmle}, that is a subset of the original
methytmle input object which has been reduced include only the relevant CpG
sites. This subset is then used to estimate VIMs.
}
\description{
Reduces the \code{methytmle} object by way of a \code{limma} model with
empirical Bayes shrinkage to include only the CpG sites below the preset
p-value cutoff.
}
