% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{methyheat}
\alias{methyheat}
\title{Heatmap for methytmle objects}
\usage{
methyheat(x, ..., n_sites = 25, type = c("raw", "ic"))
}
\arguments{
\item{x}{Object of class \code{methytmle} as produced by an appropriate call
to \code{methyvim}.}

\item{...}{Additional arguments passed \code{superheat}. This argument is
used only if \code{type} below is set to the heatmap option. Please
consult the documentation of the \code{superheat} package for options
that may be appropriately set.}

\item{n_sites}{Numeric indicating the number of CpG sites to be shown in the
plot. This argument is only used if \code{type} is set to the heatmap
option. If the number of sites analyzed is greater than this cutoff,
sites to be displayed are chosen by ranking sites based on the raw
(marginal) p-values.}

\item{type}{Whether to plot the original data (M-values or Beta-values) for
the set of top CpG sites or to plot the measurements after having
moved these into influence curve space. The latter uses the fact that
the supported parameters have asymptotically linear representations to
obtain a rotation of the data into an alternative space.}
}
\description{
Heatmap for methytmle objects
}
\examples{
library(methyvimData)
suppressMessages(library(SummarizedExperiment))
data(grsExample)
var_int <- colData(grsExample)[, 1]
methyvim_out_ate <- suppressWarnings(
 methyvim(data_grs = grsExample, sites_comp = 1, var_int = var_int,
          vim = "ate", type = "Mval", filter = "limma", filter_cutoff = 0.05,
          parallel = FALSE, tmle_type = "sl"
         )
)
methyheat(methyvim_out_ate, type = "raw")

}
