% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyvim.R
\name{methyvim}
\alias{methyvim}
\title{Differential Methylation Statistics with Variable Importance Measures}
\usage{
methyvim(data_grs, var_int = 1, vim = c("ATE", "NPVI"), type = c("Beta",
  "Mval"), filter = c("limma", "npvi", "adaptest"), filter_cutoff = 0.05,
  window_bp = 1000, corr_max = 0.35, obs_per_covar = 15,
  parallel = TRUE, return_ic = FALSE, shrink_ic = FALSE,
  tmle_type = c("glm", "super_learning"), tmle_args = list(family =
  "binomial", g_lib = c("SL.mean", "SL.glm", "SL.randomForest"), Q_lib =
  c("SL.mean", "SL.glm"), npvi_cutoff = 0.25, npvi_descr = NULL))
}
\arguments{
\item{data_grs}{An object of class \code{minfi::GenomicRatioSet}, containing
standard data structures associated with DNA Methylation experiments.
Consult the documentation for \code{minfi} to construct such objects.}

\item{var_int}{Numeric indicating the column index of the variable of
interest, whether exposure or outcome. If argument \code{vim} is set
to the ATE, then the variable of interest is treated as an exposure;
it is treated as an outcome if this is set to be the NPVI.}

\item{vim}{Character indicating the variable importance measure to be used
in the estimation procedure. Currently supported options are the ATE
for discretized exposures and NPVI for continuous exposures. ATE is
the appropriate choice when the underlying scientific question is of
the effect of an exposure on methylation, while NPVI is the parameter
of choice when the effect of methylation on an outcome is sought.}

\item{type}{Character indicating the particular measure of DNA methylation to
be used as the observed data in the estimation procedure, either Beta
values or M-values. The data are accessed via \code{minfi::getBeta} or
\code{minfi::getM}.}

\item{filter}{Character indicating...
...}

\item{filter_cutoff}{Numeric indicating...
...}

\item{window_bp}{Numeric indicating...
...}

\item{corr_max}{Numeric indicating...
...}

\item{obs_per_covar}{Numeric indicating...
...}

\item{parallel}{Logical or Numeric indicating...
...}

\item{return_ic}{Logical indicating...
...}

\item{shrink_ic}{Logical indicating...
...}

\item{tmle_type}{Character indicating the general class of regression models
to be used in fitting the propensity score and outcome regressions.
This is generally a shorthand and is overridden by \code{tmle_args} if
that argument is changed from its default values.}

\item{tmle_args}{List giving several key arguments to be passed to one of
\code{tmle::tmle} or \code{tmle.npvi::tmle.npvi}, depending on the
particular variable importance measure specified. This overrides
\code{tmle_type}, which itself provides sensible defaults. Consider
changing this away from default settings only if you have sufficient
experience with theory and software for targeted learning. For more
information, consider consulting the documentation of the \code{tmle}
and \code{tmle.npvi} packages.}
}
\value{
An object of class \code{methytmle}, with all unique slots filled in,
        in particular, including indices of CpG sites that pass screening,
        cluster of neighboring CpG sites, and a matrix of the results of the
        estimation procedure performed for the given variable importance
        measure. Optionally, estimates of the propensity score and outcome
        regressions, as well as the original data rotated into influence
        curve space may be returned, if so requested.
}
\description{
Computes the Targeted Minimum Loss-Based Estimate of a specified statistical
target parameter, formally defined within models from causal inference. The
variable importance measures currently supported are the Average Treatment
Effect (ATE) and a Nonparametric Variable Importance Measure (NPVI, formally
defined by Chambaz, Neuvial, and van der Laan <doi:10.1214/12-EJS703>).
}
