% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_npvi.R
\name{methyvim_npvi}
\alias{methyvim_npvi}
\title{Differential Methylation with a Nonparametric Variable Importance Measure}
\usage{
methyvim_npvi(methy_tmle, catch_inputs_npvi)
}
\arguments{
\item{methy_tmle}{An object of class \code{methytmle}...
...}

\item{catch_inputs_npvi}{List...
...}
}
\description{
Computes the Targeted Minimum Loss-Based Estimate of a Nonparametric
Variable Importance Measure (NPVI) of a continous exposure, treating DNA
methylation as said exposure (X) and the indicated variable of interest
as an outcome (Y), using the neighbors of a given CpG site as the adjustment
set (W).
}
