% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{methyvolc}
\alias{methyvolc}
\title{Volcano plot for methytmle objects}
\usage{
methyvolc(x, param_bound = 2, pval_bound = 0.2)
}
\arguments{
\item{x}{Object of class \code{methytmle} as produced by an appropriate call
to \code{methyvim}.}

\item{param_bound}{Numeric for a threshold indicating the magnitude of the
size of the effect considered to be interesting. This is used to
assign groupings and colors to individual CpG sites. This argument is
only used when \code{type} is set to the volcano plot option.}

\item{pval_bound}{Numeric for a threshold indicating the magnitude of
p-values deemed to be interesting. This is used to assign groupings
and colors to individual CpG sites. This argument is only used when
\code{type} is set to the volcano plot option.}
}
\description{
Volcano plot for methytmle objects
}
\examples{
library(methyvimData)
suppressMessages(library(SummarizedExperiment))
data(grsExample)
var_int <- colData(grsExample)[, 1]
methyvim_out_ate <- suppressWarnings(
 methyvim(data_grs = grsExample, sites_comp = 1, var_int = var_int,
          vim = "ate", type = "Mval", filter = "limma", filter_cutoff = 0.05,
          parallel = FALSE, tmle_type = "sl"
         )
)
methyvolc(methyvim_out_ate)

}
