% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_npvi.R
\name{npvi_screen}
\alias{npvi_screen}
\title{Screening procedure using a TMLE for nonparametric variable importance}
\usage{
npvi_screen()
}
\description{
Screening procedure using a TMLE for nonparametric variable importance
}
