% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.methytmle}
\alias{plot.methytmle}
\title{Common plots for methytmle objects}
\usage{
\method{plot}{methytmle}(x, ..., type = c("heat", "volcano", "pvals"))
}
\arguments{
\item{x}{object of class \code{biotmle} as produced by an appropriate call to
\code{biomarkertmle}}

\item{...}{additional arguments passed \code{plot} as necessary}

\item{type}{character describing whether to provide a plot of unadjusted or
adjusted p-values (adjustment performed via Benjamini-Hochberg)}
}
\value{
object of class \code{ggplot} containing one of several types of
        plots: a heatmap of the top 50 CpG sites, a volcano plot displaying
        the parameter estimate against the log-transformed raw p-values, or
        side-by-side histograms of the raw and corrected p-values.
}
\description{
Several types of plots providing...
}
