% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_parallel}
\alias{set_parallel}
\title{Parallelization with Futures and BiocParallel}
\usage{
set_parallel(parallel = c(TRUE, FALSE), future_param = NULL,
  bppar_type = NULL)
}
\arguments{
\item{parallel}{Logical indicating whether parallelization ought to be used.
Parallelization is invoked via a combination of \code{BiocParallel}
and \code{future}. If \code{TRUE} the default method uses multiprocess
evaluation, though other \code{future::plan}s may be specified using
an optional argument. If \code{FALSE}, sequential computation is used
and a warning message is issued.}

\item{future_param}{Character (if not \code{NULL}) specifying a particular
parallelization approach to be used. For a list of the options, see
the documentation for \code{future::plan}. If the previous argument
(\code{parallel}) is set to \code{FALSE}, this argument is ignored and
sequential computation is invoked via \code{future::sequential}.}

\item{bppar_type}{Character specifying the type of backend to be used with
the parallelization invoked by \code{BiocParallel}. Consult the manual
page for \code{BiocParallel::BiocParallelParam} for possible types and
descriptions on their appropriate uses. The default for this argument
is \code{NULL}, which silently uses \code{BiocParallel::DoparParam}.}
}
\value{
Nothing. This function is designed to be called for its side-effect
        of registering a parallel backend (for \code{BiocParallel}) and/or
        \code{future::plan}, making parallel computation a trivial process.
}
\description{
Easily set up a suitable parallelization scheme using the various options
provided in \code{BiocParallel} and packages of the \code{future} ecosystem.
INTERNAL USE ONLY.
}
\keyword{internal}
