% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_ic.R
\name{shrink_ic}
\alias{shrink_ic}
\title{Shrink influence curve estimates}
\usage{
shrink_ic()
}
\description{
Use moderated statistical tests (empirical Bayes methods) to shrink the
influence curve estimates of a target parameter
}
