% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inputs}
\alias{check_inputs}
\title{Check Function Inputs}
\usage{
check_inputs(data, var, filter_cutoff, window, corr, obs_per_covar, parallel,
  future, bppar, tmle_type, tmle_args)
}
\arguments{
\item{data}{...}

\item{var}{...}

\item{filter_cutoff}{...}

\item{window}{...}

\item{corr}{...}

\item{obs_per_covar}{...}

\item{parallel}{...}

\item{future}{...}

\item{bppar}{...}

\item{tmle_type}{...}

\item{tmle_args}{...}
}
\value{
Options to be passed to various \code{tmle_*} functions.
}
\description{
Check the input values of function parameters for errors.
}
