% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma.R
\name{limma_screen}
\alias{limma_screen}
\title{Screening procedure based on LIMMA}
\usage{
limma_screen(methytmle, var_int, type, cutoff = 0.05)
}
\arguments{
\item{methytmle}{An object of class \code{methytmle}.}

\item{var_int}{Numeric indicating the column index of the variable of
interest, whether exposure or outcome. If argument \code{vim} is set
to the ATE, then the variable of interest is treated as an exposure;
it is treated as an outcome if this is set to be the NPVI.}

\item{type}{Character indicating the particular measure of DNA methylation to
be used as the observed data in the estimation procedure, either Beta
values or M-values. The data are accessed via \code{minfi::getBeta} or
\code{minfi::getM}.}

\item{cutoff}{Numeric indicating the p-value cutoff that defines which
sites pass through the \code{filter}.}
}
\value{
An object of class \code{methytmle}, that is a subset of the original
methytmle input object which has been reduced include only the relevant CpG
sites. This subset is then used to estimate VIMs.
}
\description{
Screens the \code{methytmle} object by way of a \code{limma} model with
empirical Bayes shrinkage for CpG sites below the pre-defined p-value cutoff.
pre-defined p-value cutoff.
}
\details{
## or something like this? ##

Reduces the \code{methytmle} object by way of a \code{limma} model with
empirical Bayes shrinkage to include only the CpG sites below the
pre-defined p-value cutoff.
}
