% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{methyheat}
\alias{methyheat}
\title{Heatmap for methytmle objects}
\usage{
methyheat(x, ..., n_sites = 25, type = "raw")
}
\arguments{
\item{x}{Object of class \code{methytmle} as produced by an appropriate call
to \code{methyvim}.}

\item{...}{Additional arguments passed to \code{superheat}. Consult the
documentation of the \code{superheat} package for a list of options.}

\item{n_sites}{Numeric indicating the number of CpG sites to be shown in the
plot. If the number of sites analyzed is greater than this cutoff,
sites to be displayed are chosen by ranking sites based on their raw
(marginal) p-values.}

\item{type}{Whether to plot the original data (M-values or Beta-values) for
the set of top CpG sites or to plot the measurements after applying a
transformation into influence curve space (with respect to the target
parameter of interest). The latter uses the fact that the parameters
have asymptotically linear representations to obtain a rotation of the
raw data into an alternative space; moreover, in this setting, the
heatmap reduces to visualizing a supervised clustering procedure.}
}
\value{
Nothing. This function is called for its side-effect of outputting a
        heatmap to the graphics device. The heatmap is constructed using the
        \code{superheat} package.
}
\description{
Heatmap for methytmle objects
}
\examples{
suppressMessages(library(SummarizedExperiment))
library(methyvimData)
data(grsExample)
var_int <- as.numeric(colData(grsExample)[, 1])
methyvim_out_ate <- suppressWarnings(
 methyvim(data_grs = grsExample, sites_comp = 25, var_int = var_int,
          vim = "ate", type = "Mval", filter = "limma", filter_cutoff = 0.1,
          parallel = FALSE, tmle_type = "glm"
         )
)
methyheat(methyvim_out_ate, type = "raw")
}
