% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methytmle-class.R
\docType{class}
\name{methytmle-class}
\alias{methytmle-class}
\alias{.methytmle}
\alias{param}
\alias{vim}
\title{Constructor for class methytmle}
\usage{
param(object)

vim(object)
}
\arguments{
\item{object}{S4 object of class \code{methytmle}.}
}
\value{
\code{methytmle} object, subclassed from \code{GenomicRatioSet}.
}
\description{
Constructor for class methytmle

Accessor for Parameter Information

Accessor for Variable Importance Measure Information
}
\examples{
library(methyvimData)
suppressMessages(library(SummarizedExperiment))
data(grsExample)
# cast the GenomicRatioSet to class methytmle
methy_tmle <- .methytmle(grsExample)
methy_tmle
}
\keyword{internal}
