% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methytmle.R
\docType{class}
\name{methytmle-class}
\alias{methytmle-class}
\alias{.methytmle}
\title{Constructor for class methytmle}
\value{
\code{methytmle} object, subclassed from \code{GenomicRatioSet}.
}
\description{
Constructor for class methytmle
}
\examples{
library(methyvimData)
suppressMessages(library(SummarizedExperiment))
data(grsExample)
# cast the GenomicRatioSet to class methytmle
methy_tmle <- .methytmle(grsExample)
}
