% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyvim.R
\name{methyvim}
\alias{methyvim}
\title{Differential Methylation Statistics with Variable Importance Measures}
\usage{
methyvim(data_grs, var_int, vim = c("ate", "rr", "npvi"),
  type = c("Beta", "Mval"), filter = c("limma"),
  filter_cutoff = 0.05, window_bp = 1000, corr_max = 0.75,
  obs_per_covar = 20, sites_comp = NULL, parallel = TRUE,
  future_param = NULL, bppar_type = NULL, return_ic = FALSE,
  shrink_ic = FALSE, tmle_type = c("glm", "sl"),
  tmle_args = list(family = "binomial", g_lib = c("SL.mean", "SL.glm",
  "SL.bayesglm", "SL.gam"), Q_lib = c("SL.mean", "SL.glm", "SL.gam",
  "SL.earth"), npvi_cutoff = 0.25, npvi_descr = NULL),
  tmle_backend = c("tmle", "drtmle", "tmle.npvi"))
}
\arguments{
\item{data_grs}{An object of class \code{minfi::GenomicRatioSet}, containing
standard data structures associated with DNA Methylation experiments.
Consult the documentation for \code{minfi} to construct such objects.}

\item{var_int}{A \code{numeric} vector containing subject-level measurements
of the variable of interest. The length of this vector must match the
number of subjects exactly. If argument \code{vim} is set to "ate" or "rr",
then the variable of interest is treated as an exposure, and the variable
must be binary in such cases. If setting \code{vim} to target parameters
assessing continuous treatment effects, then the variable need not be binary
of course.}

\item{vim}{Character indicating the variable importance measure to be used in
the estimation procedure. Currently supported options are the ATE for
discretized exposures and NPVI for continuous exposures. ATE and RR are the
appropriate choices when the underlying scientific question is of the effect
of an exposure on methylation, while NPVI (and other continuous treatment
parameters) ought to be used when the effect of methylation on an outcome is
sought.}

\item{type}{Character indicating the particular measure of DNA methylation to
be used as the observed data in the estimation procedure, either Beta values
or M-values. The data are accessed via \code{minfi::getBeta} or
\code{minfi::getM}.}

\item{filter}{Character indicating the model to be implemented when screening
the \code{data_grs} object for CpG sites. The only currently supported
option is "limma". Contributions for other methods are welcome.}

\item{filter_cutoff}{Numeric indicating the p-value cutoff that defines which
sites pass through the \code{filter}.}

\item{window_bp}{Numeric indicating the maximum genomic distance (in base
pairs) between two sites for them to be considered neighboring sites.}

\item{corr_max}{Numeric indicating the maximum correlation that a neighboring
site can have with the target site.}

\item{obs_per_covar}{Numeric indicating the number of observations needed for
for covariate included in W for downstream analysis. This ensures the data
is sufficient to control for the covariates.}

\item{sites_comp}{A \code{numeric} indicating the maximum number of sites for
which a variable importance measure is to be estimated post-screening. This
is not typically useful in scientific settings, but may be useful when a
large number of CpG sites pass the initial screening phase.}

\item{parallel}{Logical indicating whether parallelization ought to be used.
See the documentation of \code{set_parallel} for more information, as this
argument is passed directly to that internal function.}

\item{future_param}{Character indicating the type of parallelization to be
used from the list available via the \code{future} package. See the
documentation for \code{set_parallel} for more information, as this argument
is passed directly to that internal function.}

\item{bppar_type}{Character specifying the type of backend to be used for
parallelization via \code{BiocParallel}. See the documentation for
\code{set_parallel} for more information, as this argument is passed
directly to that internal function.}

\item{return_ic}{Logical indicating whether an influence curve estimate
should be returned for each site that passed through the filter.}

\item{shrink_ic}{Logical indicating whether limma should be applied to reduce
the variance in the ic based estimates in \code{return_ic}.}

\item{tmle_type}{Character indicating the general class of regression models
to be used in fitting the propensity score and outcome regressions. This is
generally a shorthand and is overridden by \code{tmle_args} if that argument
is changed from its default values.}

\item{tmle_args}{List giving several key arguments to be passed to one of
\code{tmle::tmle} or \code{tmle.npvi::tmle.npvi}, depending on the
particular variable importance measure specified. This overrides
\code{tmle_type}, which itself provides sensible defaults. Consider changing
this away from default settings only if you have sufficient experience with
software and the underlying theory for Targeted Learning. For more
information, consider consulting the documentation of the \code{tmle} and
\code{tmle.npvi} packages.}

\item{tmle_backend}{A \code{character} indicating the package to be used in
the estimation procedure. The user should only set this parameter if they
have sufficient familiarity with the backend packages used for estimation.
Current choices include \code{tmle}, \code{drtmle}, and \code{tmle.npvi}.}
}
\value{
An object of class \code{methytmle}, with all unique slots filled in,
 in particular, including indices of CpG sites that pass screening, cluster
 of neighboring CpG sites, and a matrix of the results of the estimation
 procedure performed for the given variable importance measure. Optionally,
 estimates of the propensity score and outcome regressions, as well as the
 original data rotated into influence curve space may be returned, if so
 requested.
}
\description{
Computes the Targeted Minimum Loss-Based Estimate of a specified statistical
target parameter, formally defined within models from causal inference. The
variable importance measures currently supported are the Average Treatment
Effect (ATE) and a Nonparametric Variable Importance Measure (NPVI, formally
defined by Chambaz, Neuvial, and van der Laan <doi:10.1214/12-EJS703>).
}
\examples{
library(methyvimData)
suppressMessages(library(SummarizedExperiment))
data(grsExample)
var_int <- colData(grsExample)[, 1]
# TMLE procedure for the ATE parameter over M-values with Limma filtering
methyvim_out_ate <- suppressWarnings(
  methyvim(
    data_grs = grsExample, sites_comp = 1, var_int = var_int,
    vim = "ate", type = "Mval", filter = "limma", filter_cutoff = 0.05,
    parallel = FALSE, tmle_type = "sl"
  )
)
}
