% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_ate.R
\name{methyvim_ate}
\alias{methyvim_ate}
\title{Differential Methylation with the Average Treatment Effect}
\usage{
methyvim_ate(target_site, methytmle_screened, var_of_interest,
  type = c("Beta", "Mval"), corr = 0.8, obs_per_covar = 20,
  g_lib = c("SL.mean", "SL.glm"), Q_lib = c("SL.mean", "SL.glm"),
  family = c("gaussian", "binomial"), return_ic = FALSE)
}
\arguments{
\item{target_site}{Numeric ...}

\item{methytmle_screened}{An object of class \code{methytmle}...}

\item{var_of_interest}{...}

\item{type}{Character ...}

\item{corr}{Numeric ...}

\item{obs_per_covar}{Numeric ...}

\item{g_lib}{Character or vector of characters...}

\item{Q_lib}{Character or vector of characters...}

\item{family}{Character ...}

\item{return_ic}{Logical ...}
}
\description{
Computes the Targeted Minimum Loss-Based Estimate of the Average Treatment
Effect (ATE), treating DNA methylation as an outcome (Y) and the indicated
variable of interest (which ought to be binarized) as a treatment/exposure
(A), using the neighbors of a given CpG site as the adjustment set (W).
}
