% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_ate.R, R/tmle_npvi.R
\name{methyvim_ate}
\alias{methyvim_ate}
\alias{methyvim_ate}
\title{Differential Methylation with the Average Treatment Effect}
\usage{
methyvim_ate(methy_tmle, catch_inputs_ate)

methyvim_ate(methy_tmle, catch_inputs_ate)
}
\arguments{
\item{methy_tmle}{An object of class \code{methytmle}...
...}

\item{catch_inputs_ate}{List...
...}

\item{target_site}{Numeric ...}

\item{methy_tmle_screened}{An object of class \code{methytmle}...}

\item{i_site}{Numeric ...}

\item{type}{Character ...}

\item{corr}{Numeric ...}

\item{obs_per_covar}{Numeric ...}

\item{g_lib}{Character or vector of characters...}

\item{Q_lib}{Character or vector of characters...}

\item{family}{Character ...}

\item{return_ic}{Logical ...}
}
\description{
Computes the Targeted Minimum Loss-Based Estimate of the Average Treatment
Effect (ATE), treating DNA methylation as an outcome (Y) and the indicated
variable of interest (which ought to be binarized) as a treatment/exposure
(A), using the neighbors of a given CpG site as the adjustment set (W).

Computes the Targeted Minimum Loss-Based Estimate of a Nonparametric
Variable Importance Measure (NPVI) of a continous exposure, treating DNA
methylation as said exposure (X) and the indicated variable of interest
as an outcome (Y), using the neighbors of a given CpG site as the adjustment
set (W).
}
