% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{methyvolc}
\alias{methyvolc}
\title{Volcano plot for methytmle objects}
\usage{
methyvolc(x, param_bound = 2, pval_bound = 0.2)
}
\arguments{
\item{x}{Object of class \code{methytmle} as produced by an appropriate call
to \code{methyvim}.}

\item{param_bound}{Numeric for a threshold indicating the magnitude of the
size of the effect considered to be interesting. This is used to assign
groupings and colors to individual CpG sites.}

\item{pval_bound}{Numeric for a threshold indicating the magnitude of
p-values deemed to be interesting. This is used to assign groupings and
colors to individual CpG sites.}
}
\value{
Object of class \code{ggplot} containing a volcano plot of the
 estimated effect size on the x-axis and the -log10(p-value) on the y-axis.
 The volcano plot is used to detect possibly false positive cases, where a
 test statistic is significant due to low variance.
}
\description{
Volcano plot for methytmle objects
}
\examples{
suppressMessages(library(SummarizedExperiment))
library(methyvimData)
data(grsExample)
var_int <- as.numeric(colData(grsExample)[, 1])
methyvim_out_ate <- suppressWarnings(
 methyvim(data_grs = grsExample, sites_comp = 25, var_int = var_int,
          vim = "ate", type = "Mval", filter = "limma", filter_cutoff = 0.1,
          parallel = FALSE, tmle_type = "glm"
         )
)
methyvolc(methyvim_out_ate)
}
