% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.methytmle}
\alias{plot.methytmle}
\title{Plot p-values of methytmle objects}
\usage{
\method{plot}{methytmle}(x, ..., type = "both")
}
\arguments{
\item{x}{Object of class \code{methytmle} as produced by an appropriate call
to \code{methyvim}.}

\item{...}{Additional arguments passed \code{plot} as necessary.}

\item{type}{The type of plot to build: one of side-by-side histograms (type
"both") comparing raw p-values to FDR-adjusted p-values (using the
FDR-MSA correction) or either of these two histogram separately.}
}
\description{
Plot p-values of methytmle objects
}
\examples{
library(methyvimData)
suppressMessages(library(SummarizedExperiment))
data(grsExample)
var_int <- colData(grsExample)[, 1]
methyvim_out_ate <- suppressWarnings(
 methyvim(data_grs = grsExample, sites_comp = 1, var_int = var_int,
          vim = "ate", type = "Mval", filter = "limma", filter_cutoff = 0.05,
          parallel = FALSE, tmle_type = "sl"
         )
)
plot(methyvim_out_ate)

}
