% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.methytmle}
\alias{plot.methytmle}
\title{Plot p-values of methytmle objects}
\usage{
\method{plot}{methytmle}(x, ..., type = "both")
}
\arguments{
\item{x}{Object of class \code{methytmle} as produced by an appropriate call
to \code{methyvim}.}

\item{...}{Additional arguments passed \code{plot} as necessary.}

\item{type}{The type of plot to build: one of side-by-side histograms (type
"both") comparing raw p-values to FDR-adjusted p-values (using the FDR-MSA
correction) or either of these two histogram separately. Set this argument
to "raw_pvals" for a histogram of the raw p-values, and to "fdr_pvals" for
a histogram of the FDR-corrected p-values.}
}
\value{
Object of class \code{ggplot} containing a histogram or side-by-side
 histograms of the raw (marginal) and corrected p-values, with the latter
 computed automatically using the method of Tuglus and van der Laan.
}
\description{
Plot p-values of methytmle objects
}
\examples{
suppressMessages(library(SummarizedExperiment))
library(methyvimData)
data(grsExample)
var_int <- as.numeric(colData(grsExample)[, 1])
# TMLE procedure for the ATE parameter over M-values with Limma filtering
methyvim_out_ate <- suppressWarnings(
  methyvim(
    data_grs = grsExample, sites_comp = 25, var_int = var_int,
    vim = "ate", type = "Mval", filter = "limma", filter_cutoff = 0.1,
    parallel = FALSE, tmle_type = "glm"
  )
)
plot(methyvim_out_ate)
}
