% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_parallel}
\alias{set_parallel}
\title{Set up parallelization}
\usage{
set_parallel(parallel)
}
\arguments{
\item{parallel}{Numeric or Logical indicating either the number of cores to
use in parallelized computation or whether parallelization ought to be
used at all. If \code{TRUE}, all available cores are found via a call
to \code{parallel::detectCores} and used. If \code{FALSE}, computation
is performed sequentially (with single core) and a warning is issued.}
}
\description{
Set up parallelization
}
